model_api_deployment/requirements.txt
flask==3.0.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
psutil==5.9.8